+++ lib/modern_player.dart
@@ -1,26 +1,5 @@
 library modern_player;
 
-export 'src/modern_player.dart' show ModernPlayer;
-export 'src/others/modern_players_enums.dart'
-    show
-        ModernPlayerSubtitleSourceType,
-        ModernPlayerAudioSourceType,
-        ModernPlayerSourceType;
-export 'src/modern_player_options.dart'
-    show
-        ModernPlayerVideo,
-        ModernPlayerOptions,
-        ModernPlayerVideoData,
-        ModernPlayerControlsOptions,
-        ModernPlayerDefaultSelectionOptions,
-        ModernPlayerThemeOptions,
-        ModernPlayerProgressSliderTheme,
-        ModernPlayerCustomActionButton,
-        ModernPlayerSubtitleOptions,
-        ModernPlayerAudioTrackOptions,
-        ModernPlayerToastSliderThemeOption,
-        ModernPlayerTranslationOptions,
-        ModernPlayerCallbackOptions,
-        DefaultSelectorCustom,
-        DefaultSelectorLabel,
-        DefaultSelectorOff;
+export 'src/modern_player.dart';
+export 'src/modern_player_options.dart';
+export 'src/others/modern_players_enums.dart';